fix consensus trace parsing

+++ Code/itf/src/consensus.rs
@@ -1,3 +1,4 @@
+use num_bigint::BigInt;
 use serde::Deserialize;
 use std::collections::HashMap;
 
@@ -6,8 +7,8 @@ use crate::deserializers as de;
 pub type Address = String;
 pub type Value = String;
 pub type Step = String;
-pub type Round = i64;
-pub type Height = i64;
+pub type Round = BigInt;
+pub type Height = BigInt;
 
 #[derive(Clone, Debug, Deserialize)]
 pub enum Timeout {
@@ -132,9 +133,9 @@ impl Proposal {
     pub fn is_empty(&self) -> bool {
         self.src.is_empty()
             && self.proposal.is_empty()
-            && self.height == -1
-            && self.round == -1
-            && self.valid_round == -1
+            && self.height == BigInt::from(-1)
+            && self.round == BigInt::from(-1)
+            && self.valid_round == BigInt::from(-1)
     }
 }
 
@@ -152,8 +153,8 @@ impl VoteMessage {
     pub fn is_empty(&self) -> bool {
         self.src.is_empty()
             && self.id.is_empty()
-            && self.height == -1
-            && self.round == -1
+            && self.height == BigInt::from(-1)
+            && self.round == BigInt::from(-1)
             && self.step.is_empty()
     }
 }
